+++ capi/autd3capi/src/modulation/mod.rs
@@ -4,7 +4,7 @@
  * Created Date: 23/08/2023
  * Author: Shun Suzuki
  * -----
- * Last Modified: 06/12/2023
+ * Last Modified: 08/12/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -38,4 +38,19 @@ pub unsafe extern "C" fn AUTDModulationIntoDatagram(m: ModulationPtr) -> Datagra
 #[must_use]
 pub unsafe extern "C" fn AUTDModulationSize(m: ModulationPtr) -> ResultI32 {
     Box::from_raw(m.0 as *mut Box<M>).len().into()
+}
+
+#[repr(u8)]
+pub enum SamplingMode {
+    ExactFrequency = 0,
+    SizeOptimized = 1,
+}
+
+impl From<SamplingMode> for autd3::modulation::SamplingMode {
+    fn from(mode: SamplingMode) -> Self {
+        match mode {
+            SamplingMode::ExactFrequency => autd3::modulation::SamplingMode::ExactFrequency,
+            SamplingMode::SizeOptimized => autd3::modulation::SamplingMode::SizeOptimized,
+        }
+    }
 }