+++ cpp/include/autd3.hpp
@@ -3,7 +3,7 @@
 // Created Date: 29/05/2023
 // Author: Shun Suzuki
 // -----
-// Last Modified: 06/12/2023
+// Last Modified: 08/12/2023
 // Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 // -----
 // Copyright (c) 2023 Shun Suzuki. All rights reserved.
@@ -102,5 +102,9 @@ using internal::Controller;
 using internal::ControllerBuilder;
 
 using internal::native_methods::TimerStrategy;
+
+namespace modulation {
+using internal::native_methods::SamplingMode;
+}
 
 }  // namespace autd3