+++ CHANGELOG.md
@@ -1,5 +1,7 @@
-# 18.0.2
+# 19.0.0
 
+- Add `with_mode` to `modulation::Sine` and `modulation::Square`
+  - Fix [#231](https://github.com/shinolab/autd3/issues/231): Make modulation::Sine more flex configurable
 - Fix [#238](https://github.com/shinolab/autd3/issues/238): [C#] Simulator.WithServerIp causes crash
 
 # 18.0.1